+++ plugins/extract/detect/det1.npy
