+++ plugins/extract/detect/det3.npy
